Add element-wise vector and matrix operations
Also moved the dot product to the new protocols defined for these
operations. The reason separate protocols are needed is because of the
details of Swift’s method resolution. When a function is explicitly
defined in a protocol, the most specific implementation is used, while
if the function is just provided in an extension, the most *general*
implementation is used. This posed a problem for the operators, as
MatrixProtocol conforms to Collection, which is where the generalized
vector operations are defined as a conditional extension.